refactor: split translator into two classes
