+++ lib/night_reader.rb
@@ -4,7 +4,6 @@
 # files to plain text, assuming a 0 represents an empty space and a . represents
 # a raised dot in each cell.
 
-require './lib/translator'
 require './lib/file_writer'
 
 file_writer = FileWriter.new(ARGV[0], ARGV[1])
